users can add and delete from their pantry

+++ models/pantry_items.js
@@ -2,7 +2,12 @@ const _db     = require('./connection');
 
 module.exports = {
   getPantryItems(req,res,next) {
-    _db.any(`SELECT ingredient_name FROM pantry_items;`)
+    let uID = parseInt(req.params.id)
+    _db.any(`SELECT ingredient_name
+            FROM users
+            JOIN pantry_items
+            ON pantry_items.user_reference=users.user_id
+            WHERE users.user_id=$1;`, [uID])
        .then( pantry_item => {
         res.rows = pantry_item;
         console.log(pantry_item)
@@ -12,13 +17,16 @@ module.exports = {
         console.error('Error', error)
        })
   },
+
   addPantryItem(req,res,next) {
     console.log('=====', req.body)
+    let uID = parseInt(req.params.id)
+    console.log(uID)
     _db.one(
       `INSERT INTO
-      pantry_items (ingredient_name)
-      VALUES ($/item/)
-      returning *;`, req.body
+      pantry_items (user_reference, ingredient_name)
+      VALUES ($1, $2)
+      returning *;`, [uID, req.body.item]
     )
     .then(pantry_items => {
       console.log('Added pantry_items successfully');
@@ -49,10 +57,11 @@ module.exports = {
   },
 
   deletePantryItem(req,res,next) {
-    //const iID = Number.parseInt(req.params.item_id)
+    let uID = parseInt(req.params.id)
     _db.none(
       `DELETE FROM pantry_items
-      WHERE ingredient_name = $1;`, [req.body.item]
+      WHERE ingredient_name = $1
+      AND user_reference = $2;`, [req.body.item, uID]
     )
     .then(() => {
       console.log('Deleted pantry_items successfully');
